api/requirements.txt
azure-functions
requests
lxml